Simplify ClientPacketsHandler
