--- src/main/java/net/minestom/server/network/packet/client/handler/ClientLoginPacketsHandler.java
@@ -1,15 +1,0 @@
-package net.minestom.server.network.packet.client.handler;
-
-import net.minestom.server.network.packet.client.login.EncryptionResponsePacket;
-import net.minestom.server.network.packet.client.login.LoginPluginResponsePacket;
-import net.minestom.server.network.packet.client.login.LoginStartPacket;
-
-public class ClientLoginPacketsHandler extends ClientPacketsHandler {
-
-    public ClientLoginPacketsHandler() {
-        register(0, LoginStartPacket::new);
-        register(1, EncryptionResponsePacket::new);
-        register(2, LoginPluginResponsePacket::new);
-    }
-
-}
